# web/openapi.yaml
openapi: 3.0.3
info:
  title: Xray Checker API
  description: API for monitoring Xray proxies. Every `/api/v1/admin/*` operation requires explicitly configured HTTP Basic Auth credentials.
  version: "1.4"
servers:
  - url: /
security:
  - basicAuth: []
...
paths:
...
  /api/v1/admin/subscription/sources:
    get:
      summary: List subscription sources managed from the panel
      description: >-
        Environment sources are listed read-only beside them. Every URL is masked: a subscription URL carries the
        access token of the subscription itself, so it is never returned in full.
      tags: [Admin]
      responses:
        '200':
          description: Subscription sources
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionSources'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Add or update a subscription source
      description: >-
        Omit `id` to add, supply it to update. On update an empty `url` keeps the stored one, because the API only
        ever shows a masked URL; a `url` that still carries the mask is treated the same way, so echoing a listed
        row back cannot destroy the stored URL. An HWID is generated when the chosen client profile needs one and
        none is given; the value then stays fixed, since the remote panel ties a device slot to it. Which URLs are
        fetched changes on the next subscription refresh, but `mode`, `silent` and `unlisted` apply at once — they
        describe how nodes are watched, not what to fetch.
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionSourceRequest'
      responses:
        '200':
          description: Updated source list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionSources'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Remove a subscription source
      tags: [Admin]
      parameters:
        - in: query
          name: id
          schema: {type: string}
          required: true
      responses:
        '200':
          description: Updated source list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionSources'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message (only present on failure)
    APIErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Proxy not found"
...
    SubscriptionSources:
      type: object
      required: [sources, environmentSources, profiles]
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionSource'
        environmentSources:
          type: array
          description: Masked URLs of the sources supplied by SUBSCRIPTION_URL. They cannot be edited through this API.
          items: {type: string}
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionClientProfile'
        modes:
          type: array
          description: The selectable observation modes, so the panel renders the list rather than hard-coding it.
          items:
            $ref: '#/components/schemas/ObservationMode'
    ObservationMode:
      type: object
      properties:
        id:
          type: string
          enum: [full, availability, paused]
        label: {type: string}
        description: {type: string}
    SubscriptionSource:
      type: object
      required: [id, url, enabled, profile]
      properties:
        id: {type: string}
        url:
          type: string
          description: Masked. Scheme and host stay; the path is reduced to its first and last characters.
        name: {type: string}
        enabled: {type: boolean}
        profile:
          type: string
          enum: [checker, happ, incy, custom]
        userAgent: {type: string}
        hwid:
          type: string
          description: >-
            Returned in full. It is not a credential but the identifier the remote panel stores against this
            subscription, and it names the device slot the source occupies.
        deviceOs: {type: string}
        osVersion: {type: string}
        deviceModel: {type: string}
        mode:
          type: string
          enum: [full, availability, paused]
          description: >-
            How this source's nodes are watched. `full` is the deployment's own treatment. `availability` keeps the
            checks and the alerts but drops the source out of scheduled speed tests. `paused` still probes — a probe
            is what everything else rests on — but draws no verdict from it: no downtime, no incident, no alert, no
            speed test. A node an operator paused by hand stays paused whatever the source says.
        silent:
          type: boolean
          description: Telegram says nothing about this source. What is measured does not change.
        unlisted:
          type: boolean
          description: >-
            The source's nodes stay off the status page, out of Prometheus and without their own /config endpoint.
        createdAt: {type: string}
    SubscriptionClientProfile:
      type: object
      description: What a profile sends, so the panel can show the headers rather than only a label.
      properties:
        id: {type: string}
        label: {type: string}
        userAgent: {type: string}
        deviceOs: {type: string}
        osVersion: {type: string}
        deviceModel: {type: string}
        sendsHwid: {type: boolean}
    SubscriptionSourceRequest:
      type: object
      properties:
        id:
          type: string
          description: Absent or empty to add a new source.
        url:
          type: string
          description: Required when adding. On update an empty value keeps the stored URL.
        name: {type: string}
        enabled: {type: boolean}
        profile:
          type: string
          enum: [checker, happ, incy, custom]
          description: >-
            checker keeps the checker's own identity, which the operator's own response rules match on. happ and
            incy impersonate those iOS clients, which is what a third-party panel with an HWID device limit
            expects. custom requires an explicit userAgent.
        userAgent: {type: string}
        hwid:
          type: string
          description: 10-64 characters of latin letters, digits, '=' or '-'. Generated when empty and the profile needs one.
        deviceOs: {type: string}
        osVersion: {type: string}
        deviceModel: {type: string}
        mode:
          type: string
          enum: [full, availability, paused]
          description: Empty or unknown reads as `full`. Applied at once rather than on the next refresh.
        silent: {type: boolean}
        unlisted: {type: boolean}
...
  responses:
    Unauthorized:
      description: HTTP Basic Auth credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
...